Moves Heatmap files in proper android folder
