+++ android/lib/src/main/java/com/airbnb/android/react/maps/AirMapHeatmapManager.java
@@ -1,4 +1,4 @@
-package com.AirMaps;
+package com.airbnb.android.react.maps;
 
 import android.content.Context;
 import android.graphics.Color;
